transition phase for 'sangria.parser.SourceMapper'
