+++ modules/core/src/main/scala/sangria/parser/package.scala
@@ -1,0 +1,7 @@
+package sangria
+
+package object parser {
+
+  @deprecated("please use `sangria.ast.SourceMapper`", "3.0.0")
+  type SourceMapper = sangria.ast.SourceMapper
+}
